requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
Pillow
aiofiles
